function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai==1.12.0
tiktoken==0.6.0
opencensus-ext-azure==1.1.13
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
azure-search-documents==11.4.0
azure-ai-documentintelligence==1.0.0b1
